updated MemoryMaps
